XGE-19007: Events not generated for outbound messages when body from is not in lower case.
Jilter writes accepted event info for each mailbox in json file. The 'key' is the mailbox address which will always be in lower case.
Python producer reads the accepted event info from this file, updates it with 's3' file path and sends MH event.
While reading the accepted event info from the json file, for a mailbox, it should look it up in lowercase.