Avoiding policy reads in outbound flow.
We read policy for two purpose in outbound flow.
i) To identify sender
ii) To get customer id to see if split_by_recipient is enabled.

Both can be obtained from message history accepted event.

The change is behing a flag which is stored in EFS. Using 'xgemail.get.metadata.from.msghistory.py', the flag can be enabled for a customer, for a submit server ip or globally.