SMTPUTF8 handling.
If we received a message in submit server with 'SMTPUTF8' option, then while delivering also we need use 'SMPTUTF8' option in 'MAIL' command.
There does not seem to be a way to configure postfix to send this information to Jilter. So during delivery, checking if destination server supports 'SMTPUTF8' option.
If destination server supports it, we always send the 'SMTPUTF8' option.